+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased] - TBD
 ### Added
+* Added support for b2_set_bucket_notification_rules and b2_get_bucket_notification_rules
 * Added B2Json `fromJson` methods that take a `java.io.Reader` as input for JSON
 * Updated B2Json `fromJson` methods to utilize a BufferedReader when deserializing JSON for performance improvement
 * Added support for event notification rules